rag-document-qa/rag-document-qa/requirements.txt
# Core Dependencies
langchain
openai
python-dotenv

# Document Processing  
pypdf

# Simple Vector Store (instead of FAISS)
chromadb

# API & Frontend
fastapi
uvicorn
streamlit

# Basic Utils
tqdm
